More cleanup of XMSS headers
Merge xmss_privatekey.h and xmss_publickey.h -> xmss.h
